+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+### Added
+- `Cell.get_polygons` return a copy of a cell’s polygons including references, with the possibility of filtering by layer and datatype
+
 ## 0.7.0 - 2021-08-02
 ### Added
 - `contour` function
@@ -12,7 +16,7 @@
 - Holes in boolean results could lead to incorrect geometry in specific cases
 - Bug in boolean operations resulting in self-intersecting polygons
 - Bug in boolean operations with clockwise-oriented polygons
-- Unsupported reccords found when loading a library generate a warning, not an error.
+- Unsupported records found when loading a library generate a warning, not an error.
 ### Changed
 - `inside` has changed to use the better interfaces: grouping has been removed, scaling is not necessary, and short-circuit is implemented in `all_inside` and `any_inside`
 
@@ -27,7 +31,7 @@
 ### Changed
 - `Reference.cell` is now writable
 ### Fixed
-- Bug in Array::insert not incerasing the array count
+- Bug in Array::insert not increasing the array count
 
 ## 0.5.0 - 2021-06-11
 ### Added